Merged in vasu-display-time (pull request #8)
Update display time for artifacts

The display time previously was the time that the GUI received the
artifact. That makes it difficult to know when in the robot's path
various artifacts were detected. This commit fixes that by calculating
the DARPA time at the original detection time for an artifact. This also
modifies the display slightly.

Approved-by: Bob DeBortoli <debortor@oregonstate.edu>
Approved-by: Dan Abad <dabad@andrew.cmu.edu>